% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rox_fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{It creates a map of the LATITUDE and LONGITUD of the state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{an integer of the state code}

\item{year}{An integer of the year}
}
\value{
it returns a map of the LATITUDE and LONGITUD of the state when an observation occurs in the dataset
}
\description{
It creates a map of the LATITUDE and LONGITUD of the state
}
\details{
In case where there is not such state.num it returns a message of invalid State number. Also when LONGITUD > 900 then it is assign as NA.
		The same when LATITUDE>90
}
\examples{
fars_map_state(1,2013)
}

