% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rox_fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Data Input}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{the name of the file which the data are to be read from.}
}
\value{
returns a data frame table if the file exists, throws an error if the file
  doesn't exist. If the filename does not contain an absolute path, the filename
  is relative to the current working directory, \code{getwd()}.
}
\description{
Reads a file in table format and creates a data frame table from it.
}
\examples{
\dontrun{
  fars_read("accident_2013.csv.bz2")
}

}

