% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rox_fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Load several year of FARS data files}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a vector of integer or strings that indicated the year of the FARS data files}
}
\value{
a list of FARS data for the years in the parameter \code{years}
}
\description{
Load several year of FARS data files
}
\examples{
  \dontrun{
    fars_read_years(c(2013, 2014))
    fars_read_years(c("2013", "2014"))
  }

}

