% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rox_fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{It rbind the different tbl_df}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector of integers}
}
\value{
it takes as an input a vector of integers which are the years and it returns tbl_df object
			which is a summary of the number of observations by year and month
}
\description{
It rbind the different tbl_df
}
\details{
By using the tidyr package it spreads the years to columns so the rows are the Months and the columns the Years
}
\examples{
fars_summarize_years(c(2013,2014,2015))
}

