% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rox_fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{make a file name
creates a file name of the form accident_%d.csv.bz2 where %d is an integer}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An integer value}
}
\value{
This returns a file name of the format accident_%d.csv.bz2 where %d is an integer
        thus this is a character object
}
\description{
make a file name
creates a file name of the form accident_%d.csv.bz2 where %d is an integer
}
\examples{
make_filename(2013)



}

