% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rox_fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Draws the location of accidents in given state for a given year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{An integer that represents the state number}

\item{year}{An integer or string}
}
\value{
plots a map of the state with the accidents as dots on the map.
It throws an error if \code{"state.num"} is not invalid. If there hasn't been
any accident in that year, it produces a message.
}
\description{
Draws the location of accidents in given state for a given year
}
\examples{
  \dontrun{
    fars_map_state(48, 2013)
    fars_map_state(48, 2014)
    fars_map_state(6, 2013)
  }
}

