% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rox_fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{read csv}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A string of the filename}
}
\value{
This function takes as an input the string of the filename which is in the working directory
	   and returns a tbl_df object using the dplyr package
}
\description{
reads a filename from the working directory
}
\details{
if there is no file such a filename in the working directory it returns a message that that filename does not exist
}
\examples{
fars_read("accident_2013.csv.bz2")
}

