% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rox_fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{It returns only the two columns which are Month and the year}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of integers}
}
\value{
it takes as an input a vector of integers which are the years and it returns tbl_df object
}
\description{
It returns only the two columns which are Month and the year
}
\details{
In case that there is not a file in the working directory of that particular year it returns an error of "invalid year"
}
\examples{
fars_read_years(c(2013,2014,2015))
}

