% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rox_fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Monthly number of accidents in each year using the data from FARS data}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{a list of FARS data for the years in the parameter \code{years}}
}
\value{
A dataframe with monthly number of accidents in each row and years as
colunms. It uses \code{\link{fars_read_years}} function.
}
\description{
Monthly number of accidents in each year using the data from FARS data
}
\examples{
  \dontrun{
    fars_summarize_years(c(2013, 2014))
    fars_summarize_years(c("2013", "2014"))
  }
}

