% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rox_fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Generate file names for data files from the US National Highway Traffic Safety
Administration's Fatality Analysis Reporting System
\href{http://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}{Fatality Analysis Reporting System (FARS)}.}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An integer or string indicating the year that the data file generated}
}
\value{
an string as the name of the file
}
\description{
Generate file names for data files from the US National Highway Traffic Safety
Administration's Fatality Analysis Reporting System
\href{http://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}{Fatality Analysis Reporting System (FARS)}.
}
\examples{
\dontrun{
 make_filename(1992)
 make_filename("2012")
 }

}

